backend/requirements.txt
fastapi
uvicorn
requests
python-decouple
pydantic<2.0
pymongo[srv]==3.9
